requirements.txt
dash==2.17.1
dash_bootstrap_components==1.6.0
pandas==1.5.3
plotly==5.9.0
plotly==5.23.0
python-dotenv==1.0.1
SQLAlchemy==2.0.31
SQLAlchemy==1.4.49
SQLAlchemy==2.0.25
SQLAlchemy==1.4.46
